backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.44.0",
]
